Moved _matchJSON() -> api.api()
